Add trait `PrimitiveBytes` for `PrimitiveNumber::Bytes`
The direct constraints on `PrimitiveNumber::Bytes` did not render well
in documentation, as everything was listed on one line. With a new trait
this is more easily documented, and we can add new methods akin to the
`core::array` constructors `from_fn` and `repeat`.

This also adds `IndexMut<usize>` to the list of supertraits, as well as
`PartialEq` and `TryFrom` for slice references.