Specify error types where possible
We can't name a specific error type in `PrimitiveNumber: FromStr`
because it's different between floats and ints. However, we **can**
refine that in `PrimitiveFloat` and `PrimitiveInteger` to specify their
respective error types.

Similarly, `PrimitiveInteger: TryFrom<_>` is not specific because the
error could be `Infallible` or `TryFromIntError`. In the cases of
`PrimitiveSigned: TryFrom<i8>` and `PrimitiveUnsigned: TryFrom<u8>`
though, we do know they are always `Infallible` -- which is implied by
their respective `From`, but needs to be explicit.